Make Object::getArchetype() non pure virtual.
Here is why I've decided to do this:
i've tried creating a custom type T that inherits from `rfk::Object` and
storing it in as a field of some other type. This caused a parsing
error saying that T is abstract. I've noticed that `rfk::Object`'s
`getArchetype()` function is pure virtual and it's implemented
in the generated files, but since on parsing step we don't have
anything generated this causes the parser to show this error. To avoid
this issue I've added a default implementation for `getArchetype()`
that just returns a class with the name `NULL` and size of `int`. Since
all generated files implement `getArchetype()` I don't think this
commit will make any difference to other code.